Moved Nachricht to own file
